# Optimize Vulkan backend for better CPU performance and less GPU synchronization overhead.