@ExtReMLapin 2024-08-09
>2x on Geforce 4090 class GPUs for stable-code-instruct-3b-Q8_0.

Could you please give pp/s tg/s ?