@mtavenrath 2024-08-09
I have to correct myself since it Intel Xe isn't getting faster with this PR. I had a bogus measurement on Intel Xe before, probably because something was running the background. 

master:
```
Intel(R) Iris(R) Xe Graphics (i7-13800h)
llama_print_timings: prompt eval time =     539.70 ms /     9 tokens (   59.97 ms per token,    16.68 tokens per second)
llama_print_timings:        eval time =   18839.67 ms /   215 runs   (   87.63 ms per token,    11.41 tokens per second)

NVIDIA RTX 6000 Ada Generation
llama_print_timings: prompt eval time =      78.59 ms /     9 tokens (    8.73 ms per token,   114.51 tokens per second)
llama_print_timings:        eval time =    4262.02 ms /   280 runs   (   15.22 ms per token,    65.70 tokens per second)
```

PR:
```
Intel(R) Iris(R) Xe Graphics (i7-13800h)
llama_print_timings: prompt eval time =     555.27 ms /     9 tokens (   61.70 ms per token,    16.21 tokens per second)
llama_print_timings:        eval time =    7616.56 ms /    87 runs   (   87.55 ms per token,    11.42 tokens per second)

NVIDIA RTX 6000 Ada Generation
llama_print_timings: prompt eval time =      70.06 ms /     9 tokens (    7.78 ms per token,   128.45 tokens per second)
llama_print_timings:        eval time =    4514.31 ms /   576 runs   (    7.84 ms per token,   127.59 tokens per second)
```