@slaren 2024-08-09
| GPU         | Model         | Test   |   t/s master |   t/s vulkan_perf |   Speedup |
|:------------|:--------------|:-------|-------------:|------------------:|----------:|
| RTX 3090 Ti | llama 7B Q4_0 | pp512  |       976.00 |            997.27 |      1.02 |
| RTX 3090 Ti | llama 7B Q4_0 | tg128  |        48.89 |             72.25 |      1.48 |