@0cc4m 2024-08-10
Very nice work, this makes a big difference in CPU-limited scenarios. I'm reducing some of the GPU bottlenecks in #8959, so this should combine well with that.