@mtavenrath 2024-08-12
The potentially best next optimization would overlap copies & cmdbuffer creation with actual work on the GPU.

Below is a profiling of stable-code-instruct-3b-Q8_0 on my system. It takes ~2.5ms CPU time to create 5.5ms of GPU work. ggml_build_graph alone has an average execution time of 1.17ms in my example. If we could split cmdbuffer generation and execute split cmdbuffers early before full creating this time could potentially be completely eliminated resulting in ~15% more perf. On systems with slower CPUs (e.g. to save energy, thermal constraints, etc.) the benefit can be even bigger.

![image](https://github.com/user-attachments/assets/b55f6426-53b4-4889-a189-9e63e8a29b88)