@0cc4m 2024-08-12
Yes, that is true. I've been looking into options for that.

We can submit command buffers early so that the GPU can start to work while the rest of the graph is still being recorded. If that overcomes the overhead from more queue submissions it could reduce the cpu overhead.

We can also look into multithreading the command recording process, maybe in combination with the first option. That just needs careful synchronization to make sure it doesn't add more overhead than it resolves.