@mtavenrath 2024-08-12
When submitting early there is no need for multithreading as long as a single CPU core can keep up. Let's keep it simple first before adding the complexity which makes it hard to profile & debug.

`ggml_vk_build_graph` has a boolean `last_node` which can be used to stop cmdbuffer recording and start a new one upon the next call. `ggml_vk_compute_forward` handles this case quite well. I haven't figured out yet how to efficiently trigger the execution of a cmdbuffer upfront.