@0cc4m 2024-08-12
I would modify build_graph to submit the buffer to the queue once it gets closed, with a timeline semaphore to synchronize it with the next submission. On the last submission, add a fence and wait for that. The semaphore and fence have to go into some data structure to remember them across a run.

Do you want to look into it? If not, I'll try it when I find some time.