@mtavenrath 2024-08-21
@0cc4m I've pushed a draft for the overlapped cmdbuffer creation & execution in pr #9118. I didn't touch the semaphore handling... through it works fine on my system. I cannot tell when I'll have time to take care of the semaphores as well.